requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
aiosqlite
python-dotenv
httpx
greenlet
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
python-multipart